remove whole-process swapping
Long before Unix supported paging, it used process swapping.  While this was
ok with the PDP-11/20's 64kB address spaces, it does not work as well today
when address spaces can easily be hundreds of GB.  The process swapping code
is there and in theory it runs when we get *extremely* low on memory.  In
practice, paging keeps the system well above the swapping limit.  Swapping
only ever runs when other parts of the memory management code are doing
terrible things and doesn't actually help anything.  The complexity this
code introduces doesn't justify the little benefit it provides.

You can check the number of swapout/swapin events with kstats:

$ kstat -p ::vm:swapin ::vm:swapout

This is a change inspired by Bredan Gregg's SCaLE 12x keynote.

illumos bug #6583