bmake: build with bmake
Now that we've bootstrapped, we can clean the bmake build up a bit.
bsd.after-import.mk turned out not to work so well so we just use an
(almost) unmodified Makefile from upstream bmake and ship the requisite
configuration files so that bootstrap is not required.

This commit also improves illumos.mk, loosely based on SunOS.mk and
Generic.mk.