libm: fix lround PIC_SETUP SIGSEGV
We can't use a generic local label 0: in the PIC_SETUP macro since the macro
can be used by code that already has a 0: label.  This does not raise an
assembler error.  Instead, the assembler uses whichever 0: label it finds
first.  This may mean that instead of jumping to where the code intended,
it'll jump into the middle of PIC_SETUP.