config: convert ENABLE_SMB_PRINTING to config
Replace the ENABLE_SMB_PRINTING env file toggle with a config file entry.

This commit introduces Makefile.cfgparam.{sparc,intel} at the top level of
the repository (intended for bmake) and and same named files in usr/src
(intended for dmake).

As a gross hack, Makefile.cfgparam (in both locations) is a symlink to
Makefile.cfgparam.intel.  While we could generate the symlink on the fly,
eventually we should just replace the whole thing with Makefile.cfgparam
being generated by the build system.