vm: PVN_GETPAGE_{SZ,NUM} are misnamed and unnecessarily complicated
There is really no reason to not allow 8 pages all the time.  With the
current logic, we get the following:

Assuming 4kB pages (x86):

    _SZ  = ptob(8) /* 32kB */
    _NUM = 8

Assuming 8kB pages (sparc):

    _SZ  = ptob(8) /* 64kB */
    _NUM = 8

We'd have to deal with 16kB base pages in order for the _NUM #define to not
be 8 (it'd be 4 in that case).  So, in the spirit of simplicity, let's just
always grab 8 pages as there are no interesting systems with 16kB+ base pages.

Finally, the defines are poorly named.