netinet: remove ugly union hack
This hack doesn't actually accomplish anything since gcc is smart enough to
emit sane instructions regardless of the size of the value.  For statically
initialized globals and static locals, the value is put into the .data section
at compile time.  For automatic locals the value is constructed at run time.
For example, gcc turns:

	struct in6_addr x = IN6ADDR_ANY_INIT;

into (on amd64):

    test+0xc:  48 c7 45 f0 00 00  movq   $0x0,-0x10(%rbp)
               00 00
    test+0x14: 48 c7 45 f8 00 00  movq   $0x0,-0x8(%rbp)
               00 00

or (on i386):

    test+0x9:  c7 45 e8 00 00 00  movl   $0x0,-0x18(%ebp)
               00
    test+0x10: c7 45 ec 00 00 00  movl   $0x0,-0x14(%ebp)
               00
    test+0x17: c7 45 f0 00 00 00  movl   $0x0,-0x10(%ebp)
               00
    test+0x1e: c7 45 f4 00 00 00  movl   $0x0,-0xc(%ebp)
               00