pkg: remove system/library/iconv/utf-8 dependencies
this is a temporary measure that allows us to not depend on the iconv
libraries shipped in OI. not having the iconv libraries installed means
iconv() will always fail, but for the time being that's acceptable.
in the future we should build everything necessary from the same repo as
we build iconv from (ie. this one), but we're not quite there yet.