tcp: remove _strong_iss tunable
Instead, use random ISS on incoming connections and follow RFC6528 for
outgoing connections. This logic is borrowed from OpenBSD -- the
initial sequence numbers don't need to be increasing across connections
except for the case where the same connection-id tuple is still in
TIME_WAIT on the listener, so RFC6528 is only necessary for outgoing
connections.

The implementation is not that pretty: we use the listener tcp's tcp_iss
to pass the new ISS along to tcp_input_listener from the SYN received
while in TIME_WAIT. Previously this used to modify the tcp stack's
tcps_iss_incr_extra value though, so perhaps it's at least not worse.