fs: dnlc needs to fully initialize the negative_cache_vnode
The vnode initialization code is complicated because it is split between
several states due to the use of kmem cache constructors.  As a result,
any vnode that we call vn_reinit on must have already been initialized.

This fixes a regression caused by the recent merge of illumos-gate's
changes related to sdt DTrace probes in vnode hold macros.