fs: fsop_sync_by_kind() needs to handle NULL vfs_sync op
There is no dispatch function for this since we are using the vfssw's *copy*
of the vfs ops and not just a pointer to an ops structure.  Because of this
copying, fem never replaced these ops with the fshead_*() calls that
normally happened to struct vfs.  Furthermore, since we are passing a NULL
struct vfs pointer to the sync op, we can't make a dispatch function that
checks vfs_femhead - since there isn't one.