create_ramdisk: create cpio boot archives instead of ufs
The ufs boot archives are nothing but a headache.  Their generation takes a
lot of I/O, they take too much space in RAM, and their creation is quite
complex.

Note that bootadm still generates hsfs boot archives if it can (i.e., if
mkisofs is present).  If it cannot, it now falls back to cpio (created by
create_ramdisk).