control legacy uname with an environment variable
this allows us to remove the process flag and move the implementation
entirely into libc. it also makes puname(1) trivial; it may be removed
later.

this removes the sysinfo() return value mangling, and rtld is modified
to always receive the real uname information (for the $OSREL and
$OSNAME expansions). that's fine, though, because puname is supposed to
be a temporary hack anyway.