kernel: remove utsname_get() helper
It made sense to abstract away all the accesses to utsname because we used
to have two - the normal and a legacy one.  With the recent change to move
the uname switching code into libc, the kernel no longer has two utsname
structs to work with.  Therefore, we can remove the useless wrapper.