gitignore build-generated files in usr/src
it's huge and horrible thanks to dmake leaving crap everywhere, but this
is better than always having to use 'status -uno' and forgetting to add
untracked files due to that. bmake tree is *mostly* clean, but this does
expose some generated files we should deal with over there as well.