config: remove unnecessary knobs
They outlived their usefulness as a proof-of-concept.  Namely, it is
possible to make it easy to tweak the build.

There are still a handful of knobs left, but they either define an aspect of
the system (e.g., the machine name) or toggle whether or not various
debugging code is built in (e.g., CONFIG_VM_STATS).