interface_check: don't report "no versions found"
even if we happen to use gnu ld --version-script, but only for symbol
visibility (ie. no named version), this check is hit. remove it at least until
we can figure out how to avoid the false-positive (we don't really care that
the shared object has *versions*, but we care that symbol visibility is defined
with a version script and not everything is visible)