move 32-bit libs to lib/i386 subdirs & 64-bit libs to lib/
the most important exception: /lib/ld.so.1 and /lib/amd64/ld.so.1
remain, since they are special (PT_INTERP path is hardcoded in
executables).  however, this diff moves the actual files into
/usr/libexec/ld.so and /usr/libexec/i386/ld.so, leaving symlinks in the
old locations, for transition.

only libraries in paths that are part of the runtime linker
configuration are affected by this diff, ie. programs specifically
dlopen()ing or otherwise interacting with libraries in subdirectories,
such as /usr/lib/security, are mostly unmodified and still using the
amd64 subdir for 64-bit libs. they can be dealt with on a case-by-case
basis later. exceptions already converted are:
 - /usr/lib/scsi/plugins/*
 - /usr/lib/smedia

some dirs are switched to 64-bit only as part of this diff, including:
 - cmd/fm
 - cmd/passwd
 - cmd/prtdiag
 - cmd/abi
 - cmd/hal
 - cmd/policykit
 - lib/policykit