svc profiles: remove name_service/inetd_services.xml includes
these files are not shipped by any package, but are generally symlinks
created by caiman, an older version of kayak, or sysidcfg (none of which
we use any more).

right now the installer still creates generic.xml which is similarly not
part of any package. the system assumes that exists; for now that's
fine, but let's remove dependencies on the other links from generic.xml.

this commit also enables the network/dns/client service by default as
that's what the always-used (by kayak) ns_dns.xml profile did.