Snapshot of GPIO framework and Zen GPIO driver
This is a work in progress. At this point, the core of gpioadm, libxpio,
the kgpio driver, and zen_gpio driver are in reasonable shape. In
addition, there is an initial kgpio big theory statement and manual
pages to cover gpioadm(8) and the overall framework in gpio(7).

There is still some work to do:

o Cleaning up and improving the semantic errors
o Going through and finishing up the gpio_sim driver
o Writing up the main test suite in the os-tests to cover gpioadm,
  libxpio, and related.
o Finishing off a few remaining XXXs
o Reparenting the zen_gpio driver off of amdzen and onto the fch/huashan